react lazy imports for routes, header, footer and vadorToggle and dynamic import for heavy function in vadorToggle
